Issue 2252: Refactor the referrer hiding.
Fix #2252

Current solution for hiding the referer string requires readirecting literally
all requests for which the referrer was modified. This approach conflicts
with the fact that certain requests are not allowed to be redirected (i.e.
CORS preflight requests). That said, current approach breaks a lot of sites.

To refactor this we move the referrer hiding code to the earlier stage,
where navigation requests are created and started. This is a convenient
place to patch both renderer and browser initiated navigations and utilize
ContentBrowserClient interface on the browser thread.